# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Instance:
    properties:
      agent_auth:
        description: Encrypted agent authentication string
        type: string
      agent_port:
        minimum: 1
        type: integer
      check_interval:
        minimum: 1
        type: integer
      check_timeout:
        minimum: 1
        type: integer
      created_at:
        type: string
      description:
        type: string
      group:
        type: string
      host:
        type: string
      id:
        type: string
      label:
        type: string
      name:
        type: string
      updated_at:
        type: string
    required:
    - agent_port
    - check_interval
    - check_timeout
    - host
    - name
    type: object
  models.Report:
    properties:
      created_at:
        type: string
      file_path:
        description: Path to the generated report file
        type: string
      format:
        description: e.g., "CSV", "PDF", "Excel"
        type: string
      generated_at:
        type: string
      id:
        type: string
      name:
        type: string
      report_type:
        description: e.g., "instance_summary", "service_uptime"
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Monitron Backend.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Monitron Backend API
  version: "1.0"
paths:
...
  /instances:
    get:
      description: Retrieve a list of all monitoring instances
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Instance'
            type: array
        "500":
          description: 'error": "Could not retrieve instances'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all instances
      tags:
      - Instances
    post:
      consumes:
      - application/json
      description: Create a new monitoring instance
      parameters:
      - description: Instance object to be created
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/models.Instance'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Instance'
        "400":
          description: 'error": "Cannot parse JSON'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not create instance'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new instance
      tags:
      - Instances
  /instances/{id}:
    delete:
      description: Delete a monitoring instance by its ID
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: 'error": "Invalid instance ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error": "Instance not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not delete instance'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete an instance
      tags:
      - Instances
    get:
      description: Retrieve a single monitoring instance by its ID
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Instance'
        "400":
          description: 'error": "Invalid instance ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error": "Instance not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not retrieve instance'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get instance by ID
      tags:
      - Instances
    put:
      consumes:
      - application/json
      description: Update details of an existing monitoring instance by its ID
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      - description: Instance object with updated fields
        in: body
        name: instance
        required: true
        schema:
          $ref: '#/definitions/models.Instance'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Instance'
        "400":
          description: 'error": "Invalid instance ID" or "Cannot parse JSON'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error": "Instance not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not update instance'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update an existing instance
      tags:
      - Instances
...
  /reports:
    get:
      description: Retrieve a list of all reports
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Report'
            type: array
        "500":
          description: 'error": "Could not retrieve reports'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all reports
      tags:
      - Reports
    post:
      consumes:
      - application/json
      description: Create a new report entry and queue it for generation
      parameters:
      - description: Report object to be created
        in: body
        name: report
        required: true
        schema:
          $ref: '#/definitions/models.Report'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Report'
        "400":
          description: 'error": "Cannot parse JSON'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not create report'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new report
      tags:
      - Reports
  /reports/{id}:
    get:
      description: Retrieve a single report by its ID
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Report'
        "400":
          description: 'error": "Invalid report ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error": "Report not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error": "Could not retrieve report'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get report by ID
      tags:
      - Reports
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"